requirements.txt
# Core dependencies
mujoco>=3.0.0
gymnasium>=0.29.0
stable-baselines3>=2.2.0
numpy>=1.24.0
pyyaml>=6.0
tensorboard>=2.15.0

# Utilities
matplotlib>=3.8.0
tqdm>=4.66.0

# Testing
pytest>=7.4.0
